remove env module and replace to utils module for enum
