+++ src/pamiq_vrchat/__init__.py
@@ -1,12 +1,9 @@
 from importlib import metadata
 
 from . import actuators, sensors
-from .env import (
+from .utils import (
     ActionType,
     ObservationType,
-    VRChatAction,
-    VRChatEnvironment,
-    VRChatObservation,
 )
 
 __version__ = metadata.version(__name__.replace("_", "-"))
@@ -16,7 +13,4 @@ __all__ = [
     "sensors",
     "ActionType",
     "ObservationType",
-    "VRChatAction",
-    "VRChatObservation",
-    "VRChatEnvironment",
 ]
